// public-site/src/lib/schema.ts
export type Product = {
  id: number;
  title: string;
  slug: string;
  category: "mens wear" | "womens wear" | "kids wear";
  fabricDetails: string;
  moq: string;
  style: string | null;
  targetMarket: string | null;
  description: string;
  imageUrl: string | null;
  imageKey: string | null;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
};